--- src/react/custom-hooks/index.js
@@ -1,4 +1,0 @@
-import usePosts from "./usePosts";
-import useSinglePost from "./useSinglePost";
-
-export { usePosts, useSinglePost };
