+++ src/react/foundation/utils/contentful.js
