+++ src/react/foundation/utils/helpers.js
